big update
Improve search queries by stripping subtitles and parentheticals from titles, and removing &isbn=undefined from the api query when an ISBN is not found. Improve detection of old mobile layout. Put results in the new layout under the book description. Detect if nonfiction appears in the genre list and make the initial search 'libgen' if so and 'libgen/fiction' if not.